Rework scalingFontSize()
The function is simplified because now it only needs two parameters
and it sets the font-size directly on the elements. Then, the children
of the element can simply use a percent to scale with it.

Also, the font-size transition was removed from .book to make the text
scale immediately instead of lagging behind with a transition.